Field.subfields_as_dict always returns value lists
After further consideration its probably best if Field.subfields_as_dict
always returns a mapping of subfield codes and values where the values
are a list (since subfield codes can and do repeat). If the default
behavior ignores this possibility then we are encouraging pymarc users
to ignore the fact that a value they are looking for could be there.

refs #170